move IdWithNetwork type to src/react/types
