+++ src/react/api/search/clients.ts
@@ -3,7 +3,7 @@ import { makeAtomFamilyWithQuery } from 'jotai-query-toolkit';
 import { fetchSearch } from '../../../api/search/fetchers';
 import { SearchClientKeys } from './keys';
 
-import type { IdWithNetwork } from './types';
+import type { IdWithNetwork } from '../../types';
 import type { SearchErrorResult, SearchSuccessResult } from '@stacks/stacks-blockchain-api-types';
 
 export const makeSearchClientAtom = makeAtomFamilyWithQuery<
