+++ src/react/api/search/keys.ts
@@ -1,5 +1,5 @@
 import { makeQueryKey } from 'jotai-query-toolkit';
-import { IdWithNetwork } from './types';
+import { IdWithNetwork } from '../../types';
 
 export enum SearchClientKeys {
   Search = 'search/Search',
