--- src/react/api/search/types.ts
@@ -1,1 +1,0 @@
-export type IdWithNetwork = [id: string, networkUrl: string];
